Fix `goenv/go-build`
* Fix few mentions of pyenv. Guess replacing `pyenv => goenv` didn't quite work
* Improve readability
* Add debug support for `goenv-uninstall`
* Fix goenv-uninstall error messages starting with backtick character
and ending with single quote character. Now all start and end with
single quotes.
* Fix difference in behavior of `goenv-uninstall` when `-f` or `--force`
is given and when it's not.
Now it always checks if the version exists and fails with exit status 1 if it's missing.
Note that now `{before,after}_uninstall` hooks get triggered only if
version to be uninstalled exists.