[GH-30][GH-50] Add automatic management of env `GO{PATH,ROOT}`
Example usage with default settings for `goenv`:

```
> goenv install 1.12.0
...

> echo $GOROOT
/home/syndbg/.goenv/versions/1.12.0
> echo $GOPATH
/home/syndbg/go/1.12.0
```

* Env vars are not managed when version is `system`.
This prevents `goenv` being configured, but not used, to mess up with your system Golang setup.
* `GOENV_DISABLE_GOROOT` disables management of `GOROOT`.
Since this is new behavior that's a very drastic change, optinally disabling it is a must.
* `GOENV_DISABLE_GOPATH` disables management of `GOPATH`.
Separate configuration of both for whichever reason the end-user may have.
* `GOENV_GOPATH_PREFIX` is a configurable prefix used for the managed `GOPATH`.
The default is `$HOME/go`. Obviously this won't work for everyone, so it's configurable.
* Support for fish, bash, zsh, etc is added.